refactor: moved locate method of sphere to locate.py from onsphere.py
